requirements.txt
flask==3.0.3
tensorflow==2.15.0
numpy==1.26.4
Pillow==10.3.0